rename form component and move save logic to parent
