--- ui/lib/kv/addon/components/kv-secret-form.hbs
@@ -1,90 +1,0 @@
-<Toolbar>
-  <ToolbarFilters>
-    <Toggle @name="json" @status="success" @size="small" @checked={{this.showJsonView}} @onChange={{this.toggleJsonView}}>
-      <span class="has-text-grey">JSON</span>
-    </Toggle>
-  </ToolbarFilters>
-</Toolbar>
-
-{{#if this.showAlert}}
-  <Hds::Alert data-test-secret-version-alert @type="inline" @color="warning" class="has-top-bottom-margin" as |A|>
-    <A.Title>Warning</A.Title>
-    <A.Description>
-      You are creating a new version based on data from Version
-      {{@previousVersion}}. The current version for
-      <code>{{@secret.path}}</code>
-      is Version
-      {{@metadata.currentVersion}}.
-    </A.Description>
-  </Hds::Alert>
-{{/if}}
-
-<form {{on "submit" (perform this.save)}}>
-  <div class="box is-sideless is-fullwidth is-bottomless">
-    <NamespaceReminder @mode="create" @noun="secret" />
-    <MessageError @model={{@secret}} @errorMessage={{this.errorMessage}} />
-
-    {{#let (find-by "name" "path" @secret.allFields) as |attr|}}
-      {{#if @previousVersion}}
-        <ReadonlyFormField @attr={{attr}} @value={{get @secret attr.name}} />
-      {{else}}
-        <FormField
-          @attr={{attr}}
-          @model={{@secret}}
-          @modelValidations={{this.modelValidations}}
-          @onKeyUp={{this.pathValidations}}
-        />
-      {{/if}}
-    {{/let}}
-
-    <hr class="is-marginless has-background-gray-200" />
-    {{#if this.showJsonView}}
-      <JsonEditor
-        @title="{{if @previousVersion 'Version' 'Secret'}} data"
-        @value={{or (stringify @secret.secretData) this.emptyJson}}
-        @valueUpdated={{this.handleJson}}
-      />
-      {{#if (or this.modelValidations.secretData.errors this.lintingErrors)}}
-        <AlertInline @type={{if this.lintingErrors "warning" "danger"}} @paddingTop={{true}}>
-          {{#if this.modelValidations.secretData.errors}}
-            {{this.modelValidations.secretData.errors}}
-          {{else}}
-            JSON is unparsable. Fix linting errors to avoid data discrepancies.
-          {{/if}}
-        </AlertInline>
-      {{/if}}
-    {{else}}
-      <KvObjectEditor
-        class="has-top-margin-m"
-        @label="{{if @previousVersion 'Version' 'Secret'}} data"
-        @value={{@secret.secretData}}
-        @onChange={{fn (mut @secret.secretData)}}
-        @isMasked={{true}}
-      />
-    {{/if}}
-  </div>
-  <div class="box is-fullwidth is-bottomless">
-    <div class="control">
-      <button
-        type="submit"
-        class="button is-primary {{if this.save.isRunning 'is-loading'}}"
-        disabled={{this.save.isRunning}}
-        data-test-kv-secret-save
-      >
-        Save
-      </button>
-      <button
-        type="button"
-        class="button has-left-margin-s"
-        disabled={{this.save.isRunning}}
-        {{on "click" @onCancel}}
-        data-test-kv-secret-cancel
-      >
-        Cancel
-      </button>
-    </div>
-    {{#if this.invalidFormAlert}}
-      <AlertInline class="has-top-padding-s" @type="danger" @message={{this.invalidFormAlert}} @mimicRefresh={{true}} />
-    {{/if}}
-  </div>
-</form>
