+++ ui/lib/kv/addon/components/page/secret/edit.hbs
@@ -9,11 +9,63 @@
   </p.levelLeft>
 </PageHeader>
 
-<KvSecretForm
-  @previousVersion={{@previousVersion}}
-  @metadata={{@metadata}}
-  @secret={{@newVersion}}
-  {{! onSave transition to parent secret route to re-query latest version }}
-  @onSave={{transition-to "vault.cluster.secrets.backend.kv.secret"}}
-  @onCancel={{transition-to "vault.cluster.secrets.backend.kv.secret.details"}}
-/>
+<Toolbar>
+  <ToolbarFilters>
+    <Toggle @name="json" @status="success" @size="small" @checked={{this.showJsonView}} @onChange={{this.toggleJsonView}}>
+      <span class="has-text-grey">JSON</span>
+    </Toggle>
+  </ToolbarFilters>
+</Toolbar>
+
+{{#if this.showAlert}}
+  <Hds::Alert data-test-secret-version-alert @type="inline" @color="warning" class="has-top-bottom-margin" as |A|>
+    <A.Title>Warning</A.Title>
+    <A.Description>
+      You are creating a new version based on data from Version
+      {{@previousVersion}}. The current version for
+      <code>{{@secret.path}}</code>
+      is Version
+      {{@currentVersion}}.
+    </A.Description>
+  </Hds::Alert>
+{{/if}}
+
+<form {{on "submit" (perform this.save)}}>
+  <div class="box is-sideless is-fullwidth is-bottomless">
+    <NamespaceReminder @mode="create" @noun="secret" />
+    <MessageError @model={{@secret}} @errorMessage={{this.errorMessage}} />
+
+    <KvDataFields
+      @showJson={{this.showJsonView}}
+      @secret={{@secret}}
+      @modelValidations={{this.modelValidations}}
+      @pathValidations={{this.pathValidations}}
+      @previousVersion={{@previousVersion}}
+    />
+
+  </div>
+  <div class="box is-fullwidth is-bottomless">
+    <div class="control">
+      <button
+        type="submit"
+        class="button is-primary {{if this.save.isRunning 'is-loading'}}"
+        disabled={{this.save.isRunning}}
+        data-test-kv-secret-save
+      >
+        Save
+      </button>
+      <button
+        type="button"
+        class="button has-left-margin-s"
+        disabled={{this.save.isRunning}}
+        {{on "click" this.onCancel}}
+        data-test-kv-secret-cancel
+      >
+        Cancel
+      </button>
+    </div>
+    {{#if this.invalidFormAlert}}
+      <AlertInline class="has-top-padding-s" @type="danger" @message={{this.invalidFormAlert}} @mimicRefresh={{true}} />
+    {{/if}}
+  </div>
+</form>
