+++ ui/lib/kv/addon/components/page/secrets/create.hbs
@@ -9,8 +9,49 @@
   </p.levelLeft>
 </PageHeader>
 
-<KvSecretForm
-  @secret={{@secret}}
-  @onSave={{transition-to "vault.cluster.secrets.backend.kv.secret.details" @secret.path}}
-  @onCancel={{transition-to "vault.cluster.secrets.backend.kv.list"}}
-/>
+<Toolbar>
+  <ToolbarFilters>
+    <Toggle @name="json" @status="success" @size="small" @checked={{this.showJsonView}} @onChange={{this.toggleJsonView}}>
+      <span class="has-text-grey">JSON</span>
+    </Toggle>
+  </ToolbarFilters>
+</Toolbar>
+
+<form {{on "submit" (perform this.save)}}>
+  <div class="box is-sideless is-fullwidth is-bottomless">
+    <NamespaceReminder @mode="create" @noun="secret" />
+    <MessageError @model={{@secret}} @errorMessage={{this.errorMessage}} />
+
+    <KvDataFields
+      @showJson={{this.showJsonView}}
+      @secret={{@secret}}
+      @modelValidations={{this.modelValidations}}
+      @pathValidations={{this.pathValidations}}
+    />
+
+  </div>
+  <div class="box is-fullwidth is-bottomless">
+    <div class="control">
+      <button
+        type="submit"
+        class="button is-primary {{if this.save.isRunning 'is-loading'}}"
+        disabled={{this.save.isRunning}}
+        data-test-kv-secret-save
+      >
+        Save
+      </button>
+      <button
+        type="button"
+        class="button has-left-margin-s"
+        disabled={{this.save.isRunning}}
+        {{on "click" this.onCancel}}
+        data-test-kv-secret-cancel
+      >
+        Cancel
+      </button>
+    </div>
+    {{#if this.invalidFormAlert}}
+      <AlertInline class="has-top-padding-s" @type="danger" @message={{this.invalidFormAlert}} @mimicRefresh={{true}} />
+    {{/if}}
+  </div>
+</form>
