+++ ui/lib/kv/addon/templates/secret/details/edit.hbs
@@ -1,6 +1,6 @@
 <Page::Secret::Edit
-  @newVersion={{this.model.newVersion}}
+  @secret={{this.model.newVersion}}
   @previousVersion={{this.model.secret.version}}
-  @metadata={{this.model.metadata}}
+  @currentVersion={{this.model.metadata.currentVersion}}
   @breadcrumbs={{this.breadcrumbs}}
 />
